flask-backend/requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
cloudinary==1.33.0
contourpy==1.1.0
cycler==0.11.0
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.41.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
noise==1.2.2
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.6
